+++ benchmark/mnist/mnist_decay_sweep_test.go
@@ -90,7 +90,7 @@ func TestMNISTDecaySweep(t *testing.T) {
 		initWeightMax := 0.2
 		intScale := float64(1 << 20)
 
-		intThreshold := int32(threshold * intScale)
+		intThreshold := int64(threshold * intScale)
 		net := bio.NewNetwork(uint32(total), 0, intThreshold, dc.decayRate, 3)
 		net.LearningRule = bio.NoOpLearning{}
