+++ benchmark/mnist/mnist_int64_highscale_test.go
@@ -39,7 +39,7 @@ func TestMNISTInt64HighScale(t *testing.T) {
 
 	intScale := float64(1 << 30)
 
-	intThreshold := int32(threshold * intScale)
+	intThreshold := int64(threshold * intScale)
 	net := bio.NewNetwork(uint32(total), 0, intThreshold, decayRate, 3)
 	net.LearningRule = bio.NoOpLearning{}
